+++ include/services/BurstCollectorService.h
@@ -84,7 +84,6 @@ public:
     bool isRunning() const;
     bool forceCompleteSession(const std::string& sleep_day);
     bool generateSummaryForDate(const std::string& sleep_day);
-    bool reparseSession(const std::string& sleep_day);
 
     /**
      * Get last burst execution time